refactor(spec): split validate-map-errors schema walk from agent mapper
Extract TypeBox Cyclic path/schema inspection into validate-schema-walk.ts
and keep channel/data form classification plus mapValueErrors in
validate-map-errors.ts. Move TypeBox 1.x mapping characterization tests
next to the mapper; leave snapshot agent-error suites on validate.test.ts.
