+++ .changeset/spec-validate-map-errors-split.md
@@ -1,0 +1,7 @@
+---
+"@ggsvelte/spec": patch
+---
+
+<!-- markdownlint-disable MD041 -->
+
+Split tier-1 TypeBox error mapping into schema/path walk helpers and the agent SpecError mapper. Public validate() surface and error messages are unchanged.
